Commit 24
Ajout de crédits, changements mineurs.

+++ src/net/xemnias/client/IAPathManager.java
@@ -19,11 +19,11 @@ public class IAPathManager extends IAManager
 	{
 		Location loc = new Location((int)en.x, (int)en.y);
 		
-		if((int)originalLocation.x == loc.x && (int)originalLocation.y == loc.y) // si il est la position de départ
+		if((int)originalLocation.x == loc.x && (int)originalLocation.y == (int)loc.y) // si il est la position de départ
 		{
 			isMovingToFutur = true;
 		}
-		if((int)futurLocation.x == loc.x && (int)futurLocation.y == loc.y && goBack) // si on est a la position final && qu'on y retourne
+		if((int)futurLocation.x == loc.x && (int)futurLocation.y == (int)loc.y && goBack) // si on est a la position final && qu'on y retourne
 		{
 			isMovingToOriginal = true;
 
@@ -39,20 +39,20 @@ public class IAPathManager extends IAManager
 			
 			if((int)originalLocation.x < futurLocation.x && isMovingToFutur)
 			{
-				en.x+=(int) 100 * delta / 1000f;
+				en.x+=(int) ((100 * delta / 1000f));
 			}
 			if((int)originalLocation.x > futurLocation.x && isMovingToFutur)
 			{
-				en.x-=(int) 100 * delta / 1000f;
+				en.x-=(int) (100 * delta / 1000f);
 			}
 			
 			if((int)originalLocation.y < futurLocation.y && isMovingToFutur)
 			{
-				en.y+=(int) 100 * delta / 1000f;
+				en.y+=(int) (100 * delta / 1000f);
 			}
 			if((int)originalLocation.y > futurLocation.y && isMovingToFutur)
 			{
-				en.y-=(int) 100 * delta / 1000f;
+				en.y-=(int) (100 * delta / 1000f);
 			}
 			
 		}
@@ -64,20 +64,20 @@ public class IAPathManager extends IAManager
 			}
 			if((int)futurLocation.x < originalLocation.x && isMovingToOriginal)
 			{
-				en.x+=(int) 100 * delta / 1000f;
+				en.x+=(int) (100 * delta / 1000f);
 			}
 			if(futurLocation.x > originalLocation.x && isMovingToOriginal)
 			{
-				en.x-=(int) 100 * delta / 1000f;
+				en.x-=(int) (100 * delta / 1000f);
 			}
 			
 			if(futurLocation.y < originalLocation.y && isMovingToOriginal)
 			{
-				en.y+=(int) 100 * delta / 1000f;
+				en.y+=(int) (100 * delta / 1000f);
 			}
 			if(futurLocation.y > originalLocation.y && isMovingToOriginal)
 			{
-				en.y-= (int) 100 * delta / 1000f;
+				en.y-= (int) (100 * delta / 1000f);
 			}
 		}
 	}
